+++ bundles/event_admin/remote_provider/remote_provider_mqtt/gtest/src/CelixEarpmTestSuiteBaseClass.h
@@ -24,6 +24,7 @@
 #include <condition_variable>
 #include <vector>
 #include <cstdio>
+#include <algorithm>
 #include <gtest/gtest.h>
 
 #include "celix_bundle_context.h"