+++ bundles/remote_services/remote_service_admin_dfi/src/remote_service_admin_dfi.c
@@ -32,8 +32,8 @@
 
 #include <jansson.h>
 #include "json_serializer.h"
-#include "utils.h"
 #include "celix_utils.h"
+#include "celix_threads.h"
 #include "celix_ref.h"
 #include "hash_map.h"
 
@@ -1159,4 +1159,3 @@ static size_t remoteServiceAdmin_write(void *contents, size_t size, size_t nmemb
     fwrite(contents, size, nmemb, get->stream);
     return size * nmemb;
 }
-