+++ libs/framework/src/bundle_context.c
@@ -20,7 +20,6 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
-#include <utils.h>
 #include <assert.h>
 #include <unistd.h>
 #include <stdarg.h>