+++ libs/utils/src/filter.c
@@ -22,7 +22,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <utils.h>
+#include <celix_utils.h>
 
 #include "celix_convert_utils.h"
 #include "celix_err.h"
@@ -31,7 +31,6 @@
 #include "celix_stdio_cleanup.h"
 #include "celix_stdlib_cleanup.h"
 #include "celix_version.h"
-#include "filter.h"
 
 // ignoring clang-tidy recursion warnings for this file, because filter uses recursion
 // NOLINTBEGIN(misc-no-recursion)