+++ misc/experimental/bundles/config_admin/config_admin_tst/example_test2/private/src/activator.c
@@ -35,7 +35,6 @@
 #include "bundle_context.h"
 #include "celix_constants.h"
 #include "properties.h"
-#include "utils.h"
 /* celix.utils */
 #include "hash_map.h"
 /* celix.configadmin */