+++ CHANGES.md
@@ -58,6 +58,9 @@ limitations under the License.
 - linked_list.h is removed and no longer supported. Use celix_array_list.h instead.
 - ip_utils.h is removed and no longer supported.  
 - array_list.h is removed and no longer supported. Use celix_array_list.h instead.
+- utils.h is removed and no longer supported. Use celix_utils.h instead.
+- export.h is removed and no longer supported.
+- celix_utils_api.h is removed and no longer supported.
 - the celix_arrayList_add function no longer accepts a NULL value.
 - version.h and version_range.h are removed and no longer supported. Use celix_version.h and celix_version_range.h 
   instead.
@@ -80,6 +83,8 @@ limitations under the License.
 - The manifest format has been changed to JSON, and the manifest attribute names have been changed.
   The old manifest format is no longer supported.
   The old manifest attribute names are also no longer defined in `celix_constants.h`.
+- Using jar to packaging bundles is no longer support. Jar was used to ensure the MANIFEST.MF was the first
+  entry in a bundle zip file, but MANIFEST.MF is no longer used (replaced by MANIFEST.json).
 
 ## New Features